[droidian-migrations] Add droidian-102-polkit-suid
Droidian 102 upgrades polkit to version 127, that uses a
socket-activated service to do privilege escalation (awesome!).
Due to the wide variety of the devices we support, it is more than likely
that pidfd is not supported on the running kernel.

Keep setting SUID on the agent helper for this release.

Droidian 103 will likely handle this differently, and we will restore the
upstream behaviour on supported devices.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>